pyproject.toml
[project]
name = "fake-news-detection"
version = "0.1.0"
description = "Fake News Detection pipeline using DistilBERT"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Sergei Romanov" }
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "hydra-core>=1.3.0",
    "dvc[s3]>=3.0.0",
    "mlflow>=2.0.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "fire>=0.5.0",
    "lightning>=2.0.0",
    "kagglehub>=0.1.0",
    "onnx>=1.14.0",
    "onnxruntime>=1.15.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "onnxscript>=0.5.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pre-commit>=3.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fake_news_detector"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
max-line-length = 120
ignore = ['E203', 'W503']
